remove property_set_window
