--- som_gui/module/property_set_window/constants.py
@@ -1,3 +1,0 @@
-SEPERATOR_SECTION = "property_set_window"
-SEPERATOR_STATUS = "seperator_status"
-SEPERATOR = "seperator"
