+++ som_gui/module/property_window/qt/SplitterSettings.ui
