+++ som_gui/module/property_window/qt/ui_SplitterSettings.py
@@ -3,7 +3,7 @@
 ################################################################################
 ## Form generated from reading UI file 'SplitterSettings.ui'
 ##
-## Created by: Qt User Interface Compiler version 6.8.2
+## Created by: Qt User Interface Compiler version 6.9.0
 ##
 ## WARNING! All changes made in this file will be lost when recompiling UI file!
 ################################################################################
