+++ som_gui/module/settings/qt/Widget.ui
@@ -25,11 +25,14 @@
    </property>
    <item>
     <widget class="QTabWidget" name="tabWidget">
+     <property name="layoutDirection">
+      <enum>Qt::LayoutDirection::RightToLeft</enum>
+     </property>
      <property name="tabPosition">
       <enum>QTabWidget::TabPosition::West</enum>
      </property>
      <property name="currentIndex">
-      <number>0</number>
+      <number>2</number>
      </property>
      <widget class="QWidget" name="tabGeneral">
       <attribute name="title">
@@ -126,8 +129,8 @@
            <rect>
             <x>0</x>
             <y>0</y>
-            <width>98</width>
-            <height>28</height>
+            <width>856</width>
+            <height>497</height>
            </rect>
           </property>
           <attribute name="label">
@@ -153,9 +156,9 @@
        </item>
       </layout>
      </widget>
-     <widget class="QWidget" name="tabPropertySet">
+     <widget class="QWidget" name="tabProperty">
       <attribute name="title">
-       <string>PropertySets</string>
+       <string>Properties</string>
       </attribute>
       <layout class="QVBoxLayout" name="verticalLayout_11">
        <item>
@@ -185,8 +188,8 @@
            <rect>
             <x>0</x>
             <y>0</y>
-            <width>98</width>
-            <height>28</height>
+            <width>840</width>
+            <height>485</height>
            </rect>
           </property>
           <property name="contextMenuPolicy">
@@ -228,8 +231,8 @@
            <rect>
             <x>0</x>
             <y>0</y>
-            <width>98</width>
-            <height>28</height>
+            <width>856</width>
+            <height>527</height>
            </rect>
           </property>
           <attribute name="label">
