+++ som_gui/tool/property_table.py
@@ -21,7 +21,6 @@ import som_gui.core.tool
 from som_gui import tool
 from som_gui.module.main_window.ui import MainWindow
 from som_gui.module.project.constants import CLASS_REFERENCE
-from som_gui.module.property_set_window.ui import PropertySetWindow
 from som_gui.resources.icons import get_link_icon
 from som_gui.module.property_table.constants import MIME_DATA_KEY
 from som_gui.module.property_table import trigger
